requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
python-dotenv